fix(fields): null fields and rename routes
