+++ src/graphql/schemas/typeDefs/attestationTypeDefs.ts
@@ -1,6 +1,6 @@
 import {Field, ID, ObjectType} from "type-graphql";
 import {GraphQLBigInt, GraphQLJSON} from 'graphql-scalars';
-import type {Json} from "../../../types/supabase.js";
+import type {Json} from "../../../types/supabaseCaching.js";
 import {Hypercert} from "./hypercertTypeDefs.js";
 import {BasicTypeDef} from "./basicTypeDef.js";
