+++ src/graphql/schemas/utils/filters.ts
@@ -6,7 +6,7 @@ import {
     StringSearchOptions
 } from "../inputs/searchOptions.js";
 import type {WhereOptions} from "../inputs/whereOptions.js";
-import type {CachingDatabase} from "../../../types/supabase.js";
+import type {Database as CachingDatabase} from "../../../types/supabaseCaching.js";
 import {PostgrestTransformBuilder} from "@supabase/postgrest-js";
